+++ src/interfaces/index.ts
@@ -13,7 +13,7 @@ export {
   FlagToken, List, ListItem, BooleanFlag, Flag, FlagBase, OptionFlag,
   Input, EnumFlagOptions, DefaultContext, Default, Definition,
   CompletableOptionFlag, Completion, CompletionContext, FlagInput,
-  CompletableFlag,
+  CompletableFlag, OptionFlagProps,
 } from './parser'
 export {PJSON} from './pjson'
 export {Plugin, PluginOptions, Options} from './plugin'