+++ test/parser/parse.test.ts
@@ -1,9 +1,12 @@
 /* eslint-disable max-nested-callbacks */
 import {expect} from 'chai'
+import * as fs from 'fs'
 
 import {flags, parse} from '../../src/parser'
 import {Interfaces} from '../../src'
 import {URL} from 'url'
+import {directory, file} from '../../src/parser/flags'
+import * as sinon from 'sinon'
 
 const stripAnsi = require('strip-ansi')
 
@@ -436,6 +439,55 @@ See more help with --help`)
       expect(out.argv).to.deep.equal([100])
     })
 
+    it('parse with a default does not parse default', async () => {
+      const out = await parse([], {
+        flags: {foo: flags.string({parse: async input => input.toUpperCase(), default: 'baz'})},
+      })
+      expect(out.flags).to.deep.include({foo: 'baz'})
+    })
+
+    describe('parse with a default/value of another type (class)', async () => {
+      class TestClass {
+        public prop: string;
+        constructor(input: string) {
+          this.prop = input
+        }
+      }
+      it('uses default via value', async () => {
+        const out = await parse([], {
+          flags: {
+            foo: flags.build<TestClass>({
+              parse: async input => new TestClass(input),
+              default: new TestClass('baz'),
+            })(),
+          },
+        })
+        expect(out.flags.foo?.prop).to.equal('baz')
+      })
+      it('uses default via function', async () => {
+        const out = await parse([], {
+          flags: {
+            foo: flags.build<TestClass>({
+              parse: async input => new TestClass(input),
+              default: async () => new TestClass('baz'),
+            })(),
+          },
+        })
+        expect(out.flags.foo?.prop).to.equal('baz')
+      })
+      it('uses parser when value provided', async () => {
+        const out = await parse(['--foo=bar'], {
+          flags: {
+            foo: flags.build<TestClass>({
+              parse: async input => new TestClass(input),
+              default: new TestClass('baz'),
+            })(),
+          },
+        })
+        expect(out.flags.foo?.prop).to.equal('bar')
+      })
+    })
+
     // it('gets arg/flag in context', async () => {
     //   const out = await parse({
     //     args: [{ name: 'num', parse: (_, ctx) => ctx.arg.name!.toUpperCase() }],
@@ -812,15 +864,15 @@ See more help with --help`)
       try {
         await parse([], {
           flags: {
-            foo: flags.string({exactlyOne: ['bar']}),
-            bar: flags.string({char: 'b', exactlyOne: ['foo']}),
+            foo: flags.string({exactlyOne: ['bar', 'foo']}),
+            bar: flags.string({char: 'b', exactlyOne: ['bar', 'foo']}),
           },
         })
       } catch (error: any) {
         message = error.message
       }
 
-      expect(message).to.equal('Exactly one of the following must be provided: b,a,r')
+      expect(message).to.equal('Exactly one of the following must be provided: --bar, --foo')
     })
 
     it('throws if multiple are set', async () => {
@@ -961,6 +1013,126 @@ See more help with --help`)
         },
       })
       expect(out.flags.foo).to.equal(true)
+    })
+  })
+
+  describe('fs flags', () => {
+    const sandbox = sinon.createSandbox()
+    let existsStub: sinon.SinonStub
+    let statStub: sinon.SinonStub
+
+    beforeEach(() => {
+      existsStub = sandbox.stub(fs, 'existsSync')
+      statStub = sandbox.stub(fs.promises, 'stat')
+    })
+
+    afterEach(() => {
+      sandbox.restore()
+    })
+
+    describe('directory', () => {
+      const testDir = 'some/dir'
+      it('passes if dir !exists but exists:false', async () => {
+        const out = await parse([`--dir=${testDir}`], {
+          flags: {dir: directory({exists: false})},
+        })
+        expect(existsStub.callCount).to.equal(0)
+        expect(out.flags).to.deep.include({dir: testDir})
+      })
+      it('passes if dir !exists but exists not defined', async () => {
+        const out = await parse([`--dir=${testDir}`], {
+          flags: {dir: directory()},
+        })
+        expect(existsStub.callCount).to.equal(0)
+        expect(out.flags).to.deep.include({dir: testDir})
+      })
+      it('passes when dir exists', async () => {
+        existsStub.returns(true)
+        statStub.returns({isDirectory: () => true})
+        const out = await parse([`--dir=${testDir}`], {
+          flags: {dir: directory({exists: true})},
+        })
+        expect(out.flags).to.deep.include({dir: testDir})
+      })
+      it("fails when dir doesn't exist", async () => {
+        existsStub.returns(false)
+        try {
+          const out = await parse([`--dir=${testDir}`], {
+            flags: {dir: directory({exists: true})},
+          })
+          throw new Error(`Should have thrown an error ${JSON.stringify(out)}`)
+        } catch (error_) {
+          const error = error_ as Error
+          expect(error.message).to.equal(
+            `No directory found at ${testDir}`,
+          )
+        }
+      })
+      it('fails when dir exists but is not a dir', async () => {
+        existsStub.returns(true)
+        statStub.returns({isDirectory: () => false})
+        try {
+          const out = await parse([`--dir=${testDir}`], {
+            flags: {dir: directory({exists: true})},
+          })
+          throw new Error(`Should have thrown an error ${JSON.stringify(out)}`)
+        } catch (error_) {
+          const error = error_ as Error
+          expect(error.message).to.equal(
+            `${testDir} exists but is not a directory`)
+        }
+      })
+    })
+
+    describe('file', () => {
+      const testFile = 'some/file.ext'
+      it('passes if file doesn\'t exist but not exists:true', async () => {
+        const out = await parse([`--file=${testFile}`], {
+          flags: {file: file({exists: false})},
+        })
+        expect(out.flags).to.deep.include({file: testFile})
+        expect(existsStub.callCount).to.equal(0)
+      })
+      it('passes if file doesn\'t exist but not exists not defined', async () => {
+        const out = await parse([`--file=${testFile}`], {
+          flags: {file: file()},
+        })
+        expect(out.flags).to.deep.include({file: testFile})
+        expect(existsStub.callCount).to.equal(0)
+      })
+      it('passes when file exists', async () => {
+        existsStub.returns(true)
+        statStub.returns({isFile: () => true})
+        const out = await parse([`--file=${testFile}`], {
+          flags: {file: file({exists: true})},
+        })
+        expect(out.flags).to.deep.include({file: testFile})
+      })
+      it("fails when dir doesn't exist", async () => {
+        existsStub.returns(false)
+        try {
+          const out = await parse([`--file=${testFile}`], {
+            flags: {file: file({exists: true})},
+          })
+          throw new Error(`Should have thrown an error ${JSON.stringify(out)}`)
+        } catch (error_) {
+          const error = error_ as Error
+          expect(error.message).to.equal(`No file found at ${testFile}`)
+        }
+      })
+      it('fails when file exists but is not a file', async () => {
+        existsStub.returns(true)
+        statStub.returns({isFile: () => false})
+        try {
+          const out = await parse([`--file=${testFile}`], {
+            flags: {file: file({exists: true})},
+          })
+          throw new Error(`Should have thrown an error ${JSON.stringify(out)}`)
+        } catch (error_) {
+          const error = error_ as Error
+          expect(error.message).to.equal(`${testFile} exists but is not a file`)
+        }
+      })
     })
   })
 })